requirements.txt
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
oauthlib==3.2.2
python-decouple==3.8
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.5
tweepy==4.14.0
urllib3==2.0.4
